+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * https://scottplot.net/changelog/ is a formatted version of this document
 * https://scottplot.net/contributors/ shows all of ScottPlot's contributors
 
+## ScottPlot 5.0.15-beta (not yet on NuGet)
+* Ticks: Added additional styling options for axis tick labels (#3185) _Thanks @barnettben_
+
 ## ScottPlot 5.0.14-beta
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-01-10_
 * Rendering: Improved performance by limiting how often `AutoScale()` is called by the renderer (#3183) _Thanks @Smonze_